allow provider detect dynamically
